requirements.txt
pip~=21.2.4
Django~=4.0.4
sqlparse~=0.4.2
asgiref~=3.5.1
setuptools~=58.1.0
djangorestframework~=3.13.1
django-storages~=1.12.3
boto3~=1.22.9
botocore~=1.25.9
python-decouple~=3.6
